Refactoring XmEntitySpecService
