+++ src/test/java/com/icthh/xm/ms/entity/config/tenant/LocalXmEntitySpecService.java
@@ -1,7 +1,5 @@
 package com.icthh.xm.ms.entity.config.tenant;
 
-import static com.icthh.xm.ms.entity.config.TenantConfigMockConfiguration.getXmEntitySpec;
-
 import com.icthh.xm.commons.config.client.repository.TenantConfigRepository;
 import com.icthh.xm.commons.tenant.TenantContextHolder;
 import com.icthh.xm.ms.entity.config.ApplicationProperties;
@@ -9,10 +7,11 @@ import com.icthh.xm.ms.entity.config.XmEntityTenantConfigService;
 import com.icthh.xm.ms.entity.domain.spec.FunctionSpec;
 import com.icthh.xm.ms.entity.domain.spec.TypeSpec;
 import com.icthh.xm.ms.entity.security.access.DynamicPermissionCheckService;
-import com.icthh.xm.ms.entity.service.processor.DefinitionSpecProcessor;
-import com.icthh.xm.ms.entity.service.privileges.custom.EntityCustomPrivilegeService;
 import com.icthh.xm.ms.entity.service.XmEntitySpecService;
+import com.icthh.xm.ms.entity.service.privileges.custom.EntityCustomPrivilegeService;
+import com.icthh.xm.ms.entity.service.processor.DefinitionSpecProcessor;
 import com.icthh.xm.ms.entity.service.processor.FormSpecProcessor;
+import com.icthh.xm.ms.entity.service.spec.XmEntitySpecContextService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
@@ -21,6 +20,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import static com.icthh.xm.ms.entity.config.TenantConfigMockConfiguration.getXmEntitySpec;
+
 @Slf4j
 @Service
 @Primary
@@ -28,9 +29,6 @@ public class LocalXmEntitySpecService extends XmEntitySpecService {
 
     private final ApplicationProperties applicationProperties;
     private final TenantContextHolder tenantContextHolder;
-    private final EntityCustomPrivilegeService entityCustomPrivilegeService;
-    private final DynamicPermissionCheckService dynamicPermissionCheckService;
-    private final XmEntityTenantConfigService tenantConfigService;
 
     public LocalXmEntitySpecService(TenantConfigRepository tenantConfigRepository,
                                     ApplicationProperties applicationProperties,
@@ -40,14 +38,12 @@ public class LocalXmEntitySpecService extends XmEntitySpecService {
                                     XmEntityTenantConfigService tenantConfigService,
                                     DefinitionSpecProcessor definitionSpecProcessor,
                                     FormSpecProcessor formSpecProcessor) {
-        super(tenantConfigRepository, applicationProperties,
-            tenantContextHolder, List.of(entityCustomPrivilegeService), dynamicPermissionCheckService, tenantConfigService, definitionSpecProcessor, formSpecProcessor);
+        super(tenantConfigRepository, applicationProperties, tenantContextHolder,
+            new XmEntitySpecContextService(definitionSpecProcessor, formSpecProcessor, tenantConfigService),
+            List.of(entityCustomPrivilegeService), dynamicPermissionCheckService);
 
         this.applicationProperties = applicationProperties;
         this.tenantContextHolder = tenantContextHolder;
-        this.entityCustomPrivilegeService = entityCustomPrivilegeService;
-        this.dynamicPermissionCheckService = dynamicPermissionCheckService;
-        this.tenantConfigService = tenantConfigService;
     }
 
     @Override
